requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
celery-progress==0.3
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.1.0
Django==4.1.7
django-ajax-datatable==4.5.0
django-celery-results==2.5.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-modeltranslation==0.18.11
Faker==24.9.0
kombu==5.3.6
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
gunicorn
six